Move DataManager to separate module
Avoids circular import
