+++ newsfragments/XXX.bugfix
@@ -1,0 +1,1 @@
+Move ``DataManager`` to separate module to avoid circular import
